Move all users models from home to users app
